--- client/src/components/Addresses/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './Addresses';
