+++ client/src/pages/mypage/MyPage.tsx
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
 import * as S from './styles';
 import MyPageAside from '@/components/MyPage/MyPageAside';
-import Addresses from '@/components/Addresses';
+import Address from '@/components/Address';
 import { MY_PAGE_NAVIGATIONS } from '@/contstants';
 
 const MyPage = () => {
-  const [contentValue, setContentValue] = useState('addresses');
+  const [contentValue, setContentValue] = useState('address');
 
   const contentName = MY_PAGE_NAVIGATIONS.find(
     (nav) => nav.value === contentValue
@@ -14,7 +14,7 @@ const MyPage = () => {
   const renderBody = () => {
     // TODO: router Switch 사용 고민
     if (contentValue === 'orderHistroy') return <div></div>;
-    if (contentValue === 'addresses') return <Addresses />;
+    if (contentValue === 'address') return <Address />;
     if (contentValue === 'coupone') return <div></div>;
     if (contentValue === 'inqurey') return <div></div>;
     if (contentValue === 'review') return <div></div>;
