+++ client/src/pages/mypage/styles.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import Title from '@/components/Title';
+import Title from '@/components/Shared/Title';
 
 export const MyPageContainer = styled.div`
   position: relative;
